docs: update mesh demo
