+++ tensorboard/plugins/mesh/mesh_demo.py
@@ -20,8 +20,9 @@ from absl import flags
 import numpy as np
 import tensorflow as tf
 
-from tensorboard.plugins.mesh import summary as mesh_summary
+from tensorboard.plugins.mesh import summary_v2 as mesh_summary
 from tensorboard.plugins.mesh import demo_utils
+
 
 flags.DEFINE_string(
     "logdir", "/tmp/mesh_demo", "Directory to write event logs to."
@@ -30,17 +31,27 @@ flags.DEFINE_string("mesh_path", None, "Path to PLY file to visualize.")
 
 FLAGS = flags.FLAGS
 
-tf.compat.v1.disable_v2_behavior()
-
 # Max number of steps to run training with.
 _MAX_STEPS = 10
 
 
+def train_step(vertices, faces, colors, config_dict, step):
+    """Executes summary as a train step."""
+    # Change colors over time.
+    t = float(step) / _MAX_STEPS
+    transformed_colors = t * (255 - colors) + (1 - t) * colors
+    mesh_summary.mesh(
+        "mesh_color_tensor",
+        vertices=vertices,
+        faces=faces,
+        colors=transformed_colors,
+        config_dict=config_dict,
+        step=step,
+    )
+
+
 def run():
-    """Runs session with a mesh summary."""
-    # Mesh summaries only work on TensorFlow 1.x.
-    if int(tf.__version__.split(".")[0]) > 1:
-        raise ImportError("TensorFlow 1.x is required to run this demo.")
+    """Runs training steps with a mesh summary."""
     # Flag mesh_path is required.
     if FLAGS.mesh_path is None:
         raise ValueError(
@@ -57,43 +68,20 @@ def run():
     faces = np.expand_dims(faces, 0)
     colors = np.expand_dims(colors, 0)
 
-    # Create placeholders for tensors representing the mesh.
-    step = tf.placeholder(tf.int32, ())
-    vertices_tensor = tf.placeholder(tf.float32, vertices.shape)
-    faces_tensor = tf.placeholder(tf.int32, faces.shape)
-    colors_tensor = tf.placeholder(tf.int32, colors.shape)
+    # Create summary writer.
+    writer = tf.summary.create_file_writer(FLAGS.logdir)
 
-    # Change colors over time.
-    t = tf.cast(step, tf.float32) / _MAX_STEPS
-    transformed_colors = t * (255 - colors) + (1 - t) * colors
-
-    meshes_summary = mesh_summary.op(
-        "mesh_color_tensor",
-        vertices=vertices_tensor,
-        faces=faces_tensor,
-        colors=transformed_colors,
-        config_dict=config_dict,
-    )
-
-    # Create summary writer and session.
-    writer = tf.summary.FileWriter(FLAGS.logdir)
-    sess = tf.Session()
-
-    for i in range(_MAX_STEPS):
-        summary = sess.run(
-            meshes_summary,
-            feed_dict={
-                vertices_tensor: vertices,
-                faces_tensor: faces,
-                colors_tensor: colors,
-                step: i,
-            },
-        )
-        writer.add_summary(summary, global_step=i)
+    with writer.as_default():
+        for step in range(_MAX_STEPS):
+            train_step(vertices, faces, colors, config_dict, step)
 
 
 def main(unused_argv):
     print("Saving output to %s." % FLAGS.logdir)
+    print(
+        "To view results in your browser, run `tensorboard --logdir %s`"
+        % FLAGS.logdir
+    )
     run()
     print("Done. Output saved to %s." % FLAGS.logdir)
